Rename service module to apiserver
